requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.18.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
python-dotenv>=1.0.0
tenacity>=8.0.0
pyyaml>=6.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0

# ML dependencies (optional for Phase 3+)
# scikit-learn>=1.3.0
# numpy>=1.24.0
# pandas>=2.0.0
